tasks

+++ lib/tasks/pro.rake
@@ -1,14 +1,17 @@
-namespace :dev do
+namespace :pro do
 
   DEFAULT_PASSWORD = 123456
   DEFAULT_FILES_PATH = File.join(Rails.root, 'lib', 'tmp')
 
-  desc "Configura o ambiente..."
+  desc "Configura o ambiente de produção"
   task setup: :environment do
+    if Rails.env.production?
       show_spinner("Apagando BD...") { %x(rails db:drop) }
       show_spinner("Criando BD...") { %x(rails db:create) }
       show_spinner("Migrando BD...") { %x(rails db:migrate) }
       show_spinner("Cadastrando o administrador padrão...") { %x(rails dev:add_default_admin) }
+    else
+      puts "Você não está em ambiente de produção!"
     end
   end
